backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
click==8.1.8
cryptography==44.0.2
dateparser==1.2.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2023.12.25
reportlab==4.4.0
rsa==4.9.1
scikit-learn==1.4.2
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
spacy==3.7.4
SQLAlchemy==2.0.40
starlette==0.46.2
threadpoolctl==3.5.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1

# Email processing dependencies
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
emails==0.6.0

# Queue management
celery==5.3.4
redis==5.0.1
